Added add task screen to task feature
