+++ packages/timesheets/lib/features/activity/activity.dart
@@ -3,7 +3,6 @@ export 'presentation/pages/activity_page.dart';
 export 'presentation/pages/activity_router_page.dart';
 export 'presentation/pages/activity_routes.dart';
 export 'presentation/activity_ongoing_widget.dart';
-export 'presentation/activity_start_widget.dart';
 export 'presentation/activity_title_widget.dart';
 export 'presentation/activity_syncing_widget.dart';
 export 'blocs/activity_cubit.dart';
