+++ packages/timesheets/lib/features/tasks/presentation/pages/task_routes.dart
@@ -1,10 +1,16 @@
 import 'package:timesheets/features/tasks/tasks.dart';
 import 'package:timesheets/configurations/configurations.dart';
 
+import 'create_task_page.dart';
+
 const taskRoutes = [
   AutoRoute(
     path: '',
     page: TaskPage,
     initial: true,
+  ),
+  AutoRoute(
+    path: 'create_task',
+    page: CreateTaskPage,
   ),
 ];
